requirements.txt
aiofiles==0.8.0
ajsonrpc==1.2.0
anyio==3.6.1
asgiref==3.9.1
async-timeout==4.0.2
bottle==0.12.23
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
diff-match-patch==20241021
Django==5.2.5
django-crispy-forms==1.14.0
django-import-export==4.3.9
django-resized==1.0.3
django-taggit==6.1.0
h11==0.16.0
idna==3.3
ifaddr==0.2.0
marshmallow==3.17.1
packaging==21.3
pillow==11.3.0
platformio==6.1.4
psycopg==3.2.9
psycopg-binary==3.2.9
pyelftools==0.29
pyparsing==3.0.9
pyserial==3.5
python-decouple==3.8
requests==2.32.4
semantic-version==2.10.0
sniffio==1.3.0
sqlparse==0.5.0
starlette==0.47.2
tablib==3.8.0
tabulate==0.8.10
typing_extensions==4.14.1
urllib3==1.26.12
uvicorn==0.18.3
wsproto==1.1.0
zeroconf==0.39.0
